Operator page1 integration

+++ src/auth/AuthContext.tsx
@@ -1,14 +1,16 @@
 import { createContext, useContext, useState, ReactNode } from "react";
 import { toaster } from "../components/toaster";
+import { ApiService } from "../api/services";
 // import { useOperatorCart } from "../context/OperatorCartContext";
 
 interface AuthState {
   token: string | null;
   role: "Admin" | "Operator" | null;
+  // userName: string | null;
 }
 
 interface AuthContextType extends AuthState {
-  login: (token: string, role: "Admin" | "Operator") => void;
+  login: (token: string, role: "Admin" | "Operator", userName: string) => void;
   logout: () => void;
 }
 
@@ -23,19 +25,24 @@ export const AuthProvider = ({ children }: { children: ReactNode }) => {
     (localStorage.getItem("role") as "Admin" | "Operator") || null
   );
 
-  const login = (token: string, role: "Admin" | "Operator") => {
+  const login = (token: string, role: "Admin" | "Operator", userName: string) => {
     localStorage.setItem("token", token);
     localStorage.setItem("role", role);
+    localStorage.setItem("userName", userName);
     setToken(token);
     setRole(role);
   };
 
-  const logout = () => {
-    localStorage.clear();
-    // clearCart()
-    setToken(null);
-    toaster.success('Logged Out', "Success");
-    setRole(null);
+  const logout = async() => {
+    debugger
+    const response = await ApiService.logout() as any
+    if(response) {
+      localStorage.clear();
+      // clearCart()
+      setToken(null);
+      toaster.success('Logged Out', "Success");
+      setRole(null);
+    }
   };
 
   return (
